Sprint 98 DSI roles and EOI extension limit

+++ src/Sfa.Tl.Find.Provider.Api/Jobs/EmployerInterestCleanupJob.cs
@@ -21,6 +21,16 @@ public class EmployerInterestCleanupJob : IJob
     {
         try
         {
+            if ((await context.Scheduler.GetCurrentlyExecutingJobs())
+                .Any(x =>
+                    x.FireInstanceId != context.FireInstanceId
+                    && x.JobDetail.Key.Equals(context.JobDetail.Key)))
+            {
+                _logger.LogInformation("Duplicate job detected for {jobKey} - exiting immediately.",
+                    context.JobDetail.Key.Name);
+                return;
+            }
+
             await _employerInterestService.RemoveExpiredEmployerInterest();
             await _employerInterestService.NotifyExpiringEmployerInterest();
